+++ src/commands/system/users.js
@@ -1,3 +1,4 @@
+const chalk = require("chalk");
 const si = require("systeminformation");
 const ora = require("ora");
 
@@ -8,6 +9,8 @@ const Verbose = require("../../classes/Verbose");
 
 const users = async (...args) => {
   try {
+    const advanced = args.includes("-a");
+
     Verbose.startSpinner();
     const spinner = ora({ text: "Please wait..." }).start();
 
@@ -17,18 +20,15 @@ const users = async (...args) => {
     Verbose.stopSpinner();
 
     data.forEach((user) => {
-      _showKeyValue(
-        user,
-        new Map([
-          ["user", "Username"],
-          ["tty", "Terminal"],
-          ["date", "Login Date"],
-          ["time", "Login Time"],
-          ["ip", "Remote Login IP"],
-          ["command", "Last Command or Shell"],
-        ]),
-        "user"
-      );
+      console.log(chalk.bold.red(user.user));
+      console.log(`Logged in at ${user.time} on ${user.date}.\n`);
+
+      if (advanced) {
+        console.log(`Terminal: ${user.tty}`);
+        if (user.ip) console.log(`Remote Login IP: ${user.ip}`);
+
+        console.log();
+      }
     });
   } catch (err) {
     Verbose.nonFatalError();